+++ apps/backend/src/Obscura.Application/DependencyInjection.cs
@@ -3,6 +3,7 @@ using Obscura.Application.Collections;
 using Obscura.Application.Entities;
 using Obscura.Application.Jobs;
 using Obscura.Application.Media;
+using Obscura.Application.Migrations;
 using Obscura.Application.Taxonomy;
 using Obscura.Application.Videos;
 
@@ -26,6 +27,7 @@ public static class DependencyInjection
         services.AddScoped<MediaService>();
         services.AddScoped<TaxonomyService>();
         services.AddScoped<JobService>();
+        services.AddScoped<SystemMigrationService>();
 
         return services;
     }